+++ packages/agent/CHANGELOG.md
@@ -12,6 +12,18 @@
 
 ### Removed
 
+## [2026.8.30] - 2026-08-30
+
+### Breaking Changes
+
+### Added
+
+### Changed
+
+### Fixed
+
+### Removed
+
 ## [2026.8.29] - 2026-08-29
 
 ### Breaking Changes